functions/requirements.txt
# Function dependencies, for example:
# package>=version
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.9
click==8.0.3
cloudevents==1.2.0
deprecation==2.1.0
Flask==2.0.2
functions-framework==3.0.0
google-api-core==2.2.2
google-auth==2.3.3
google-cloud-core==2.2.1
google-cloud-datastore==2.4.0
googleapis-common-protos==1.53.0
grpcio==1.42.0
grpcio-status==1.42.0
gunicorn==20.1.0
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
libcst==0.3.23
MarkupSafe==2.0.1
mypy-extensions==0.4.3
packaging==21.3
proto-plus==1.19.8
protobuf==3.19.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.6
PyYAML==6.0
requests==2.26.0
rsa==4.8
shortuuid==1.0.8
six==1.16.0
typing-extensions==4.0.1
typing-inspect==0.7.1
urllib3==1.26.7
watchdog==1.0.2
Werkzeug==2.0.2